wgpu-core: Move identity code to new `identity` module.
