+++ player/src/lib.rs
@@ -15,7 +15,9 @@ use std::{borrow::Cow, fmt::Debug, fs, marker::PhantomData, path::Path};
 #[derive(Debug)]
 pub struct IdentityPassThrough<I>(PhantomData<I>);
 
-impl<I: Clone + Debug + wgc::id::TypedId> wgc::hub::IdentityHandler<I> for IdentityPassThrough<I> {
+impl<I: Clone + Debug + wgc::id::TypedId> wgc::identity::IdentityHandler<I>
+    for IdentityPassThrough<I>
+{
     type Input = I;
     fn process(&self, id: I, backend: wgt::Backend) -> I {
         let (index, epoch, _backend) = id.unzip();
@@ -26,7 +28,7 @@ impl<I: Clone + Debug + wgc::id::TypedId> wgc::hub::IdentityHandler<I> for Ident
 
 pub struct IdentityPassThroughFactory;
 
-impl<I: Clone + Debug + wgc::id::TypedId> wgc::hub::IdentityHandlerFactory<I>
+impl<I: Clone + Debug + wgc::id::TypedId> wgc::identity::IdentityHandlerFactory<I>
     for IdentityPassThroughFactory
 {
     type Filter = IdentityPassThrough<I>;
@@ -34,7 +36,7 @@ impl<I: Clone + Debug + wgc::id::TypedId> wgc::hub::IdentityHandlerFactory<I>
         IdentityPassThrough(PhantomData)
     }
 }
-impl wgc::hub::GlobalIdentityHandlerFactory for IdentityPassThroughFactory {}
+impl wgc::identity::GlobalIdentityHandlerFactory for IdentityPassThroughFactory {}
 
 pub trait GlobalPlay {
     fn encode_commands<A: wgc::hal_api::HalApi>(
@@ -47,7 +49,7 @@ pub trait GlobalPlay {
         device: wgc::id::DeviceId,
         action: trace::Action,
         dir: &Path,
-        comb_manager: &mut wgc::hub::IdentityManager,
+        comb_manager: &mut wgc::identity::IdentityManager,
     );
 }
 
@@ -151,7 +153,7 @@ impl GlobalPlay for wgc::global::Global<IdentityPassThroughFactory> {
         device: wgc::id::DeviceId,
         action: trace::Action,
         dir: &Path,
-        comb_manager: &mut wgc::hub::IdentityManager,
+        comb_manager: &mut wgc::identity::IdentityManager,
     ) {
         use wgc::device::trace::Action;
         log::info!("action {:?}", action);
