+++ wgpu-core/src/command/bundle.rs
@@ -91,8 +91,9 @@ use crate::{
     },
     error::{ErrorFormatter, PrettyError},
     hal_api::HalApi,
-    hub::{GlobalIdentityHandlerFactory, Hub, Resource, Storage, Token},
+    hub::{Hub, Resource, Storage, Token},
     id,
+    identity::GlobalIdentityHandlerFactory,
     init_tracker::{BufferInitTrackerAction, MemoryInitKind, TextureInitTrackerAction},
     pipeline::{self, PipelineFlags},
     resource,
