+++ wgpu-core/src/command/clear.rs
@@ -7,8 +7,9 @@ use crate::{
     get_lowest_common_denom,
     global::Global,
     hal_api::HalApi,
-    hub::{self, GlobalIdentityHandlerFactory, Token},
+    hub::{self, Token},
     id::{BufferId, CommandEncoderId, DeviceId, TextureId, Valid},
+    identity::GlobalIdentityHandlerFactory,
     init_tracker::{MemoryInitKind, TextureInitRange},
     resource::{Texture, TextureClearMode},
     track::{TextureSelector, TextureTracker},
