+++ wgpu-core/src/command/query.rs
@@ -6,8 +6,9 @@ use crate::{
     command::{CommandBuffer, CommandEncoderError},
     global::Global,
     hal_api::HalApi,
-    hub::{GlobalIdentityHandlerFactory, Storage, Token},
+    hub::{Storage, Token},
     id::{self, Id, TypedId},
+    identity::GlobalIdentityHandlerFactory,
     init_tracker::MemoryInitKind,
     resource::QuerySet,
     Epoch, FastHashMap, Index,
