+++ wgpu-core/src/device/queue.rs
@@ -10,8 +10,9 @@ use crate::{
     get_lowest_common_denom,
     global::Global,
     hal_api::HalApi,
-    hub::{GlobalIdentityHandlerFactory, Input, Token},
+    hub::Token,
     id,
+    identity::{GlobalIdentityHandlerFactory, Input},
     init_tracker::{has_copy_partial_init_tracker_coverage, TextureInitRange},
     resource::{BufferAccessError, BufferMapState, StagingBuffer, TextureInner},
     track, FastHashSet, SubmissionIndex,
