+++ wgpu-core/src/resource.rs
@@ -2,8 +2,9 @@ use crate::{
     device::{DeviceError, HostMap, MissingDownlevelFlags, MissingFeatures},
     global::Global,
     hal_api::HalApi,
-    hub::{GlobalIdentityHandlerFactory, Resource, Token},
+    hub::{Resource, Token},
     id::{AdapterId, DeviceId, SurfaceId, TextureId, Valid},
+    identity::GlobalIdentityHandlerFactory,
     init_tracker::{BufferInitTracker, TextureInitTracker},
     track::TextureSelector,
     validation::MissingBufferUsageError,
